Add testing for git python functions
