+++ python/mooseutils/__init__.py
@@ -10,10 +10,9 @@
 from .mooseutils import colorText, str2bool, find_moose_executable, runExe, check_configuration
 from .mooseutils import find_moose_executable_recursive, run_executable
 from .mooseutils import touch, unique_list, gold, make_chunks, camel_to_space
-from .mooseutils import text_diff, git_ls_files, git_root_dir, is_git_repo, unidiff, text_unidiff
-from .mooseutils import run_profile, list_files, git_commit, git_merge_commits, git_commit_message
-from .mooseutils import check_output, run_time
+from .mooseutils import text_diff, unidiff, text_unidiff, run_profile, list_files, check_output, run_time
 from .mooseutils import generate_filebase, recursive_update
+from .gitutils import is_git_repo, git_commit, git_commit_message, git_merge_commits, git_ls_files, git_root_dir
 from .message import mooseDebug, mooseWarning, mooseMessage, mooseError
 from .MooseException import MooseException
 from .eval_path import eval_path
